requirements.txt
openai>=1.40.0
gradio==4.44.1
python-dotenv
tiktoken
gradio
pymupdf
numpy
